restructure main.rs to allow for benchmarks
